# application/services/custom_services/module/src/models/entitymodels/modules.ts
export interface modules 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   name: { type: String },
   email: { type: String }
}